Intento 1 de crear producto

+++ app/src/main/java/com/example/socios/modelo/Producto.kt
@@ -4,6 +4,6 @@ data class Producto(
     val codigo: String,
     val nombre: String,
     val descripcion: String,
-    val precio: Double,
+    val precio: Int,
     val mail: String
 )
